+++ HISTORY.md
@@ -71,6 +71,7 @@
 ### 2022-12-01
 * blocks: don't show "result pic" option in context menu with "click-to-run" disabled
 * blocks: added configuration option for "noRingify"
+* blocks: fixed search
 
 ### 2022-11-30
 * German translation update